+++ neuronics/opensource/lib/py/tf/shannon/mutual_information_grad/mutual_information_grad.cu.cc
@@ -11,7 +11,7 @@ using GPUDevice = Eigen::GpuDevice;
 template <typename dtype>
 void MutualInformationGradFunctor<GPUDevice, dtype>::operator()(
     const GPUDevice& d,
-    const float grad,
+    const double grad,
     const dtype x,
     const dtype y
   ) {
